+++ noodles-gff/CHANGELOG.md
@@ -4,12 +4,12 @@
 
 ### Changed
 
-  * gff/record/attributes: Change underlying structure to an `IndexMap<Key,
+  * gff/record/attributes: Change underlying structure to an `IndexMap<Tag,
     Value>` ([#183]).
 
     This allows record attribute fields to have multiple values. Replace usages
-    of, e.g., `attributes.iter().find(|entry| entry.key() == key).map(|entry|
-    entry.value())` with `attributes.get(key).and_then(|values|
+    of, e.g., `attributes.iter().find(|entry| entry.key() == tag).map(|entry|
+    entry.value())` with `attributes.get(tag).and_then(|values|
     values.first())`.
 
     Values are now wrapped with `Value`, a newtype for `Vec<String>`. Replace